refactor: 🛠️ moved Stripe file to gateways
